added cones

+++ srcs/misc.c
@@ -6,7 +6,7 @@
 /*   By: wkornato <wkornato@student.42warsaw.pl>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/22 19:54:07 by wkornato          #+#    #+#             */
-/*   Updated: 2024/12/07 17:35:58 by wkornato         ###   ########.fr       */
+/*   Updated: 2024/12/31 16:31:25 by wkornato         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -36,7 +36,7 @@ int	rgb_to_int(int r, int g, int b)
 
 int	vector_to_int(t_vector vector)
 {
-	return (rgb_to_int(vector.x, vector.y, vector.z));
+	return (rgb_to_int((int)vector.x, (int)vector.y, (int)vector.z));
 }
 
 t_vector	color_to_vector(t_color color)
